Rename MusicStorage to Actions and add query to get all music memories
