+++ app/src/main/java/com/example/musicmap/screens/main/FeedFragment.java
@@ -10,7 +10,10 @@ import android.widget.ListView;
 import com.example.musicmap.R;
 import com.example.musicmap.feed.FeedAdapter;
 import com.example.musicmap.feed.MusicMemory;
+import com.example.musicmap.util.firebase.Queries;
 import com.google.firebase.firestore.GeoPoint;
+
+import org.checkerframework.checker.units.qual.A;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -22,15 +25,6 @@ public class FeedFragment extends MainFragment {
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View feedView = inflater.inflate(R.layout.fragment_feed, container, false);
         List<MusicMemory> feed = new ArrayList<>();
-
-        // TODO replace with actual feed elements
-        for (int i = 0; i < 10; i++) {
-            // CSOFF: LineLength
-            String imageUri = "https://www.agconnect.nl/sites/ag/files/2020-05/tu_eindhoven_photo_-_bart_van_overbeeke.jpg.png";
-            // CSON: LineLength
-            String song = "3B7udSGy2PfgoCniMSb523";
-            feed.add(new MusicMemory("You", new Date(), new GeoPoint(51.4486, 5.4907), imageUri, song));
-        }
 
         Activity activity = requireActivity();
         FeedAdapter feedAdapter = new FeedAdapter(activity, R.layout.single_post_layout_feed, feed);
